+++ src/main.cpp
@@ -118,7 +118,7 @@ int main() {
     
     end = std::chrono::system_clock::now();
   
-    std::chrono::duration<double> elapsed_seconds = end - start;
+    std::chrono::duration<float> elapsed_seconds = end - start;
     std::time_t end_time = std::chrono::system_clock::to_time_t(end);
   
     std::cout << "finished computation at " << std::ctime(&end_time)
